# swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  types.Host:
    properties:
      address:
        example: host.example.org
        type: string
      mac:
        example: AA:BB:CC:DD:EE:FF
        type: string
      name:
        example: my-host
        type: string
    required:
    - mac
    - name
    type: object
  types.HostStatus:
    properties:
      address:
        type: string
      error:
        type: string
      mac:
        type: string
      online:
        type: boolean
    type: object
  v1.Response:
    properties:
      reason:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: Manage known hosts and send magic packets.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: NetRouse API
  version: "1.0"
paths:
  /hosts:
    get:
      description: Fetch all known hosts
      produces:
      - application/json
      responses:
        "200":
          description: List of all known hosts
          schema:
            items:
              $ref: '#/definitions/types.Host'
            type: array
        "500":
          description: Failed to retrieve hosts from storage
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Get hosts
    put:
      consumes:
      - application/json
      description: Add a new host to the known hosts
      parameters:
      - description: New host to add
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.Host'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
        "400":
          description: Invalid MAC address or hostname
          schema:
            $ref: '#/definitions/v1.Response'
        "403":
          description: Storage is readonly
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Failed to add host
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Add new host
  /hosts/{macAddr}:
    delete:
      description: Remove a host from the list of known hosts
      parameters:
      - description: MAC address of the host
        in: path
        name: macAddr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
        "400":
          description: Invalid MAC address
          schema:
            $ref: '#/definitions/v1.Response'
        "403":
          description: Storage is readonly
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Failed to remove host
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Remove host
  /hosts/status:
    get:
      description: Get the status of all hosts with an address to check if they are
        online
      produces:
      - application/json
      responses:
        "200":
          description: List of host statuses
          schema:
            items:
              $ref: '#/definitions/types.HostStatus'
            type: array
        "500":
          description: Failed to fetch or ping hosts
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Get host status
  /wake/{macAddr}:
    get:
      description: Send a magic packet to the specified MAC address
      parameters:
      - description: MAC address of the host
        in: path
        name: macAddr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
        "400":
          description: Invalid MAC address
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Failed to send magic packet
          schema:
            $ref: '#/definitions/v1.Response'
      summary: Wake up host
produces:
- application/json
swagger: "2.0"